Adding d_fcache. Minors

+++ pub/jpeg.h
@@ -6,6 +6,9 @@
 #endif
 
 #include <stdint.h>
+#include <stdarg.h>
+#include <stdio.h>
+#include <string.h>
 
 typedef struct {
     uint16_t w, h;
@@ -14,7 +17,7 @@ typedef struct {
 } jpeg_info_t;
 
 int jpeg_init (const char *conf);
-void *jpeg_cache (const char *path, uint32_t *size);
+void *jpeg_cache (const char *path, size_t *size);
 void jpeg_release (void *p);
 int jpeg_decode (jpeg_info_t *info, void *tempbuf, void *data, uint32_t size);
 void *jpeg_2_rawpic (const char *path, void *tmpbuf, uint32_t bufsize);
